Remove Burgers and others, renumber remaining
The two Numba notebooks are not at all up-to-date and also don't contain
any content that is required to complete the course.

The Burgers notebook was never finished but can be picked up for a
separate lesson/course later (it still exists in the history of this
repo).

The odd even decoupling notebook was started as an extra resource to
explain some behavior in the solution of the cavity flow problem. Might
stick it somewhere else later.

Since we had the Numba notebook as the 14th out of 16 total notebooks,
I've shifted the leading numbers of Steps 11 and 12 down by one.